Files cleaned up, translation page logic built and working
