+++ .gitignore
@@ -1,3 +1,6 @@
 venv/
 __pycache__/
 .env
+audio_stream/*.wav
+audio_stream/*.txt
+audio_stream/clips/*.wav
