--- Nail-API/requirements.txt
@@ -1,39 +1,0 @@
-aiofiles==23.2.1
-aiohappyeyeballs==2.4.0
-aiohttp==3.10.5
-aiohttp-retry==2.8.3
-aiosignal==1.3.1
-anyio==4.3.0
-attrs==24.2.0
-blinker==1.8.2
-certifi==2024.8.30
-charset-normalizer==3.3.2
-click==8.1.7
-colorama==0.4.6
-dataclass-factory==2.16
-Flask==3.0.3
-frozenlist==1.4.1
-gunicorn==23.0.0
-idna==3.8
-iniconfig==2.0.0
-itsdangerous==2.2.0
-Jinja2==3.1.4
-MarkupSafe==2.1.5
-multidict==6.1.0
-numpy==1.26.4
-packaging==24.1
-pluggy==1.5.0
-PyAudio==0.2.14
-pydantic==1.10.18
-pydub==0.25.1
-pytest==8.3.3
-python-dotenv==1.0.1
-redis==5.0.8
-requests==2.32.3
-shazamio==0.6.0
-shazamio_core==1.0.7
-sniffio==1.3.1
-typing_extensions==4.12.2
-urllib3==2.2.2
-Werkzeug==3.0.4
-yarl==1.12.1
