--- Nail-API/templates/history.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>LyricLens</title>
-</head>
-<body>
-    <h1>Shazamio Audio Recognition- history!</h1>
-    <img src="{{ url_for('static', filename='images/l3.png') }}" style="height: 100px; width: 100px;">
-
-    <h1>
-        <a href="{{ url_for('hello_world') }}">Link to home</a>
-    </h1>
-</body>
