--- Nail-API/trans.py
@@ -1,9 +1,0 @@
-def detect(text):
-    # x = api call(text)
-    # return x
-    return None
-
-def translate(text):
-    # x = api call(text)
-    # return x
-    return None
